refactor iocage.Storage
